add if condition

+++ hw4/run.sh
@@ -1,2 +1,5 @@
-wget -O ./model/model-39500.data-00000-of-00001 "https://www.dropbox.com/s/pcblcj2ehx9vg1m/model-39500.data-00000-of-00001?dl=1"
+if ! [ -f "./model/model-39500.data-00000-of-00001" ]; then
+	wget -O ./model/model-39500.data-00000-of-00001 "https://www.dropbox.com/s/pcblcj2ehx9vg1m/model-39500.data-00000-of-00001?dl=1"
+fi
+
 python generate.py --test_path $1
